LPS-155165 Rename and remove 1 interface
